requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart
anthropic
requests
beautifulsoup4
biopython
python-pptx
openpyxl
Pillow
python-dotenv
